+++ AI_DecisionTrees/src/pl/lodz/uni/math/decisionTrees/attributes/FriOrSat.java
@@ -67,7 +67,7 @@ public class FriOrSat extends TreeAttribute {
             }
         }
         
-        return (((examplesAlternateTrue.size())/examples.size())*TreeAttribute.finalDecisionEntropy(examplesAlternateTrue))+ (((examplesAlternateFalse.size())/examples.size())*TreeAttribute.finalDecisionEntropy(examplesAlternateFalse));
+        return ((((double)examplesAlternateTrue.size())/(double)examples.size())*TreeAttribute.finalDecisionEntropy(examplesAlternateTrue))+ ((((double)examplesAlternateFalse.size())/(double)examples.size())*TreeAttribute.finalDecisionEntropy(examplesAlternateFalse));
     }
 
 }
